LPS-176921 Move to com.liferay.object.field because this enum belongs to object field
